clear one test for build_test
